removing old higher order tutorial
