+++ graft.cabal
@@ -26,7 +26,6 @@ library
       Effect.Writer
       Effect.Writer.Passthrough
       Examples.Ltl.HigherOrder
-      Examples.Ltl.HigherOrderAlternative
       Examples.Ltl.Simple
       Logic.Ltl
       Logic.Ltl.TH
